requirements.txt
PyQt5>=5.15.7
paramiko>=2.11.0
Pillow>=9.2.0